requirements.txt
docx2txt>=0.9
gradio==5.17.1
langchain==0.3.19
langchain-community==0.3.18
langchain-core==0.3.37
langchain-google-genai>=2.0.11
langchain-openai==0.3.6
langgraph==0.2.74
pandas>=2.2.3
pydantic==2.10.6
python-dotenv==1.0.1
simpleidml>=1.2.0 
